+++ packages/dd-trace/src/plugins/plugin.js
@@ -42,6 +42,7 @@ module.exports = class Plugin {
     storage.enterWith({ ...store, span })
   }
 
+  // lets you skip the creation of a span
   skip () {
     const store = storage.getStore()
     this._storeStack.push(store)
